chore: moving session manager to a repository
